Added requirements
Made it so you have to be signed in to modify create or delete games. Made it so that it will remove the buttons for those options based on if your signed in or not.

+++ delete.php
@@ -1,4 +1,11 @@
 <?php
+require_once('./utils/lib_auth.php');
+
+if (!is_logged('uID')){ 
+	header('location: index.php');
+	die();
+}
+
 require_once('./utils/JSONfunctions.php');
 deleteValue($_GET['index'],'./data/data.json');
 header("Location:index.php?delete=yes");
